chore(factor-v2): separate utils
